+++ src/main/java/com/fiveonechain/digitasset/auth/UserContext.java
@@ -10,24 +10,24 @@ import java.util.List;
  * Created by yuanshichao on 2016/11/8.
  */
 public class UserContext {
-    private final Long userId;
+    private final Integer userId;
     private final String userName;
     private final List<GrantedAuthority> authorities;
 
-    private UserContext(Long userId, String userName, List<GrantedAuthority> authorities) {
+    private UserContext(Integer userId, String userName, List<GrantedAuthority> authorities) {
         this.userId = userId;
         this.userName = userName;
         this.authorities = authorities;
     }
 
-    public static UserContext create(Long userId, String userName, List<GrantedAuthority> authorities) {
+    public static UserContext create(Integer userId, String userName, List<GrantedAuthority> authorities) {
         Assert.notNull(userId);
         Assert.notNull(userName);
         Assert.notEmpty(authorities);
         return new UserContext(userId, userName, authorities);
     }
 
-    public Long getUserId() {
+    public int getUserId() {
         return userId;
     }
